+++ RNTester/js/TextExample.android.js
@@ -15,6 +15,7 @@ var ReactNative = require('react-native');
 var {Image, StyleSheet, Text, View} = ReactNative;
 var RNTesterBlock = require('./RNTesterBlock');
 var RNTesterPage = require('./RNTesterPage');
+const TextLegend = require('./Shared/TextLegend');
 
 class Entity extends React.Component<$FlowFixMeProps> {
   render() {
@@ -86,6 +87,9 @@ class TextExample extends React.Component<{}> {
           <Text style={{padding: 10}}>
             This text is indented by 10px padding on all sides.
           </Text>
+        </RNTesterBlock>
+        <RNTesterBlock title="Text metrics legend">
+          <TextLegend />
         </RNTesterBlock>
         <RNTesterBlock title="Font Family">
           <Text style={{fontFamily: 'sans-serif'}}>Sans-Serif</Text>
